# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Megamouth",
        "contact": {},
        "license": {
            "name": "Kizuku"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/post/all": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "post"
                ],
                "summary": "投稿の全件取得",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/post/create": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "post"
                ],
                "summary": "投稿の投稿",
                "parameters": [
                    {
                        "description": "postの新規作成",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/schema.PostInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.PostOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/post/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "post"
                ],
                "summary": "post_idから投稿を取得",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.PostOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            },
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "post"
                ],
                "summary": "投稿の編集",
                "parameters": [
                    {
                        "description": "postの編集",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/schema.PostInput"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.PostOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "post"
                ],
                "summary": "投稿の削除",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.PostOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/user/create": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "userの新規作成",
                "parameters": [
                    {
                        "description": "userの新規作成",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/schema.SignUpInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.AuthOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/user/is_used/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "IDが使用済みかどうか(使用済みはtrueが返る)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.MessageOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "サインイン",
                "parameters": [
                    {
                        "description": "サインイン",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/schema.SignInInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.AuthOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "userIDからUserを返す",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.UserOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
        "schema.AuthOutput": {
            "type": "object",
            "properties": {
                "jwt": {
                    "type": "string"
                }
            }
        },
        "schema.MessageOutput": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "boolean"
                }
            }
        },
        "schema.PostInput": {
            "type": "object",
            "required": [
                "content",
                "user_id"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "schema.PostOutput": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "reaction": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.ReactionOutput"
                    }
                },
                "reply": {
                    "$ref": "#/definitions/schema.PostOutput"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "schema.ReactionOutput": {
            "type": "object",
            "properties": {
                "reaction_id": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/schema.UserOutput"
                }
            }
        },
        "schema.SignInInput": {
            "type": "object",
            "required": [
                "id",
                "password"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "schema.SignUpInput": {
            "type": "object",
            "required": [
                "id",
                "image_url",
                "password"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "schema.UserOutput": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "basic"
        }
    }
}